+++ src/server/db.js
@@ -66,6 +66,7 @@ export class Mongo {
    */
   constructor({ url, dbname, cacheSize, mockClient }) {
     if (cacheSize === undefined) cacheSize = 1000;
+    if (dbname === undefined) dbname = 'bgio';
 
     this.client = mockClient || MongoClient;
     this.url = url;
@@ -88,6 +89,28 @@ export class Mongo {
    * @param {object} store - A game state to persist.
    */
   async set(id, state) {
+    // Don't set a value if the cache has a more recent version.
+    // This can occur due a race condition.
+    //
+    // For example:
+    //
+    // A --sync--> server | DB => 0 --+
+    //                                |
+    // A <--sync-- server | DB => 0 --+
+    //
+    // B --sync--> server | DB => 0 ----+
+    //                                  |
+    // A --move--> server | DB <= 1 --+ |
+    //                                | |
+    // A <--sync-- server | DB => 1 --+ |
+    //                                  |
+    // B <--sync-- server | DB => 0 ----+
+    //
+    const cacheValue = this.cache.get(id);
+    if (cacheValue && cacheValue._stateID >= state._stateID) {
+      return;
+    }
+
     this.cache.set(id, state);
 
     const col = this.db.collection(id);
@@ -104,9 +127,9 @@ export class Mongo {
    *                     if no game is found with this id.
    */
   async get(id) {
-    const item = this.cache.get(id);
-    if (item !== undefined) {
-      return item;
+    let cacheValue = this.cache.get(id);
+    if (cacheValue !== undefined) {
+      return cacheValue;
     }
 
     const col = this.db.collection(id);
@@ -116,7 +139,27 @@ export class Mongo {
       .limit(1)
       .toArray();
 
-    this.cache.set(id, docs[0]);
+    let oldStateID = 0;
+    cacheValue = this.cache.get(id);
+    /* istanbul ignore next line */
+    if (cacheValue !== undefined) {
+      /* istanbul ignore next line */
+      oldStateID = cacheValue._stateID;
+    }
+
+    let newStateID = -1;
+    if (docs.length > 0) {
+      newStateID = docs[0]._stateID;
+    }
+
+    // Update the cache, but only if the read
+    // value is newer than the value already in it.
+    // A race condition might overwrite the
+    // cache with an older value, so we need this.
+    if (newStateID >= oldStateID) {
+      this.cache.set(id, docs[0]);
+    }
+
     return docs[0];
   }
 
@@ -126,8 +169,8 @@ export class Mongo {
    * @returns {boolean} - True if a game with this id exists.
    */
   async has(id) {
-    const item = this.cache.get(id);
-    if (item !== undefined) {
+    const cacheValue = this.cache.get(id);
+    if (cacheValue !== undefined) {
       return true;
     }