+++ src/server/index.js
@@ -9,12 +9,12 @@
 const Koa = require('koa');
 const IO = require('koa-socket');
 const Redux = require('redux');
-import { InMemory } from './db';
+import { InMemory, Mongo } from './db';
 import { createGameReducer } from '../core/reducer';
 const PING_TIMEOUT = 20 * 1e3;
 const PING_INTERVAL = 10 * 1e3;
 
-function Server({ games, db, _clientInfo, _roomInfo }) {
+export function Server({ games, db, _clientInfo, _roomInfo }) {
   const app = new Koa();
   const io = new IO({
     ioOptions: {
@@ -26,9 +26,12 @@ function Server({ games, db, _clientInfo, _roomInfo }) {
   io.attach(app);
 
   if (db === undefined) {
-    db = new InMemory();
+    if (process.env.MONGO_URI) {
+      db = new Mongo({ url: process.env.MONGO_URI });
+    } else {
+      db = new InMemory();
+    }
   }
-  db.connect();
 
   const clientInfo = _clientInfo || new Map();
   const roomInfo = _roomInfo || new Map();
@@ -86,8 +89,10 @@ function Server({ games, db, _clientInfo, _roomInfo }) {
             }
           }
 
-          db.set(gameID, store.getState());
+          await db.set(gameID, store.getState());
         }
+
+        return;
       });
 
       socket.on('sync', async (gameID, playerID, numPlayers) => {
@@ -103,6 +108,7 @@ function Server({ games, db, _clientInfo, _roomInfo }) {
         clientInfo.set(socket.id, { gameID, playerID });
 
         let state = await db.get(gameID);
+
         if (state === undefined) {
           const store = Redux.createStore(reducer);
           state = store.getState();
@@ -130,7 +136,12 @@ function Server({ games, db, _clientInfo, _roomInfo }) {
     });
   }
 
-  return app;
+  return {
+    app,
+    db,
+    run: async (port, callback) => {
+      await db.connect();
+      app.listen(port, callback);
+    },
+  };
 }
-
-export default Server;